Move defines and functions
