+++ powqutyd/files/Makefile
@@ -58,6 +58,7 @@ LD_LIBS+=-lpthread
 #LD_PATH+=-L/usr/local/lib/
 
 OBJS= src/emulator.o \
+        src/file_checks.o \
 	src/helper.o \
 	src/config.o \
 	src/uci_config.o \
